+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - `signing_key`, `signing_key_file`, `signing_ttl` attributes [#39](https://github.com/avenga/couper-vscode/pull/39)
   - `jwks_url`, `jwks_ttl` attributes, `backend` block or reference [#40](https://github.com/avenga/couper-vscode/pull/40)
   - `beta_scope_claim` [#41](https://github.com/avenga/couper-vscode/pull/41)
+  - `beta_role_claim`, `beta_role_map` [#42](https://github.com/avenga/couper-vscode/pull/42)
 
 <a name="v0.6.0"></a>